+++ tests/testthat/test-listings.R
@@ -350,3 +350,17 @@ testthat::test_that("split_into_pages_by_var works as expected", {
   )
   testthat::expect_identical(lsting, lsting_id)
 })
+
+testthat::test_that("appropriate error message returned for 'difftime' class", {
+  tmp_data <- ex_adae[1:100, ]
+  class(tmp_data$study_duration_secs) <- "difftime"
+
+  testthat::expect_error(as_listing(
+    tmp_data,
+    key_cols = c("USUBJID", "AGE"),
+    disp_cols = "study_duration_secs",
+    main_title = "title",
+    main_footer = "foot"
+  ) %>%
+    split_into_pages_by_var("SEX", page_prefix = "Patient Subset - Sex"))
+})